WIP with bugs?

+++ src/video/draw/d_edge.c
@@ -189,7 +189,8 @@ void D_DrawSurfaces(){
 			d_zistepv = s->d_zistepv;
 			d_ziorigin = s->d_ziorigin;
 
-			D_DrawSolidSurface (s, (int)s->data & 0xFF);
+			// D_DrawSolidSurface(s, (int)s->data & 0xFF);
+			D_DrawSolidSurface(s, (int)((uintptr_t)s->data & 0xFF));
 			D_DrawZSpans (s->spans);
 		}
 	}
